+++ share/classes/semester.class.php
@@ -103,15 +103,18 @@ class Semester {
      * @return mixed 
      */
     public function add(){
-        $db = DB::prepare('SELECT COUNT(id) FROM semester WHERE UPPER(semester) = UPPER(?) AND institution_id = ?');
-        $db->execute(array($this->semester, $this->institution_id));
-        if($db->fetchColumn >= 1) { 
-            return 'Diesen Lernzeitraum gibt es bereits.';
-        } else {
-            $db = DB::prepare('INSERT INTO semester (semester,description,begin,end,creation_time,creator_id,institution_id)
-                                            VALUES (?,?,?,?,NOW(),?,?)');
-            return $db->execute(array($this->semester, $this->description, $this->begin, $this->end, $this->creator_id, $this->institution_id));	
-        }   
+        global $USER;
+        if (checkCapabilities('semester:add', $USER->role_id)){
+            $db = DB::prepare('SELECT COUNT(id) FROM semester WHERE UPPER(semester) = UPPER(?) AND institution_id = ?');
+            $db->execute(array($this->semester, $this->institution_id));
+            if($db->fetchColumn >= 1) { 
+                return 'Diesen Lernzeitraum gibt es bereits.';
+            } else {
+                $db = DB::prepare('INSERT INTO semester (semester,description,begin,end,creation_time,creator_id,institution_id)
+                                                VALUES (?,?,?,?,NOW(),?,?)');
+                return $db->execute(array($this->semester, $this->description, $this->begin, $this->end, $this->creator_id, $this->institution_id));	
+            }   
+        }
     }
     
     /**
@@ -119,8 +122,11 @@ class Semester {
      * @return boolean 
      */
     public function update(){
-        $db = DB::prepare('UPDATE semester SET semester = ?, description = ?, begin = ?, end = ? WHERE id = ?');
-        return $db->execute(array($this->semester, $this->description, $this->begin, $this->end, $this->id));
+        global $USER;
+        if (checkCapabilities('semester:update', $USER->role_id)){
+            $db = DB::prepare('UPDATE semester SET semester = ?, description = ?, begin = ?, end = ? WHERE id = ?');
+            return $db->execute(array($this->semester, $this->description, $this->begin, $this->end, $this->id));
+        }
     }
     
     /**
@@ -128,22 +134,25 @@ class Semester {
      * @return boolean 
      */
     public function delete($creator_id = null){
-        if ($creator_id != null) { // if function is called by request-php --> required by checkCapabilities()
+        /*if ($creator_id != null) { // if function is called by request-php --> required by checkCapabilities()
             $user = new USER();
 
             $user->load('id', $creator_id);
             $role_id = $user->role_id;
         } else {
             $role_id = $USER->role-id;
-        } 
-        $db = DB::prepare('SELECT id FROM groups WHERE semester_id = ?');
-        $db->execute(array($this->id));           
-        $result = $db->fetchObject();
-        if ($result){
-            return false; 
-        } else {
-            $db = DB::prepare('DELETE FROM semester WHERE id = ?');
-            return $db->execute(array($this->id));
+        } */
+        global $USER;
+        if (checkCapabilities('semester:delete', $USER->role_id)){
+            $db = DB::prepare('SELECT id FROM groups WHERE semester_id = ?');
+            $db->execute(array($this->id));           
+            $result = $db->fetchObject();
+            if ($result){
+                return false; 
+            } else {
+                $db = DB::prepare('DELETE FROM semester WHERE id = ?');
+                return $db->execute(array($this->id));
+            }
         }
     }
